screenshot_verification/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 深度学习框架
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
transformers==4.35.2

# 图像处理
opencv-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.21.0
albumentations==1.3.1

# 数据处理
numpy==1.24.3
pandas==2.1.3
scipy==1.11.4

# 机器学习
scikit-learn==1.3.2
xgboost==2.0.2

# 图像分析
imagehash==4.3.1
pytesseract==0.3.10

# 元数据处理
exifread==3.0.0
piexif==1.1.3

# 可视化和分析
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# 模型可解释性
captum==0.6.0
grad-cam==1.4.0

# 超参数调优
optuna==3.5.0

# 加密和安全
cryptography==41.0.7
bcrypt==4.1.2

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1

# 缓存
redis==5.0.1

# 监控和日志
prometheus-client==0.19.0
structlog==23.2.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 开发工具
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# 其他工具
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0